+++ src/lib/appState/main/slice.ts
@@ -1,21 +1,11 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { fetchHistoryThunk } from './operations';
-import { Order } from '@/lib/types';
 
 type initialStateType = {
   shopView: 'table' | 'list';
-  page: number;
-  limit: number;
-  totalPages: number;
-  history: Order[];
 };
 
 const initialState: initialStateType = {
   shopView: 'table',
-  page: 1,
-  limit: 10,
-  totalPages: 0,
-  history: [],
 };
 
 const appStateSlice = createSlice({
@@ -25,13 +15,6 @@ const appStateSlice = createSlice({
     setShopView: (state, { payload }) => {
       state.shopView = payload;
     },
-  },
-  extraReducers: (builder) => {
-    builder.addCase(fetchHistoryThunk.fulfilled, (state, { payload }) => {
-      state.history = payload.orders;
-      state.page = payload.page;
-      state.totalPages = payload.totalPages;
-    });
   },
 });